+++ app/src/main/java/com/ichi2/apisample/helper/equality/EqualityChecker.java
@@ -1,0 +1,5 @@
+package com.ichi2.apisample.helper.equality;
+
+public interface EqualityChecker {
+    boolean areEqual(String v1, String v2);
+}
