+++ app/src/main/java/com/ichi2/apisample/helper/search/SearchExpressionMaker.java
@@ -1,0 +1,6 @@
+package com.ichi2.apisample.helper.search;
+
+public interface SearchExpressionMaker {
+    String getExpression(String value);
+    boolean isDefinitive();
+}
